Refactor and rename Add to Campaign dialog
